Rename shared files
