+++ src/booking/domain/shared/entity-like.ts
