+++ src/booking/domain/shared/value-object.ts
